add search results display
