+++ src/main/resources/view/SearchStatusView.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.text.Text?>
+<?import javafx.scene.layout.StackPane?>
+
+<StackPane minWidth="300" maxWidth="Infinity" prefHeight="24" styleClass="searchStatus"
+           xmlns:fx="http://javafx.com/fxml/1" fx:controller="seedu.todo.ui.view.SearchStatusView">
+    <HBox fx:id="searchStatusContent" minWidth="0" maxWidth="Infinity" spacing="2.0" StackPane.alignment="CENTER_LEFT">
+        <Text styleClass="searchLabel">Searching for:</Text>
+        <Text fx:id="searchTerm" styleClass="searchTerm" text="search terms" />
+    </HBox>
+
+    <Text fx:id="searchCount" styleClass="searchCount" text="0 results found" StackPane.alignment="CENTER_RIGHT" />
+</StackPane>
